Added collected_static and imagecontent to .gitignore.
There's no reason to version control collected_static because it's
never edited and can be automatically regenereated by running python
manage.py collectstatic.

imagecontent is where the default ImageField contenttype is saving its
uploads. Once again, these are all testing files and don't need to be
version control. In the future, I'd like this to be a subdir of a media
directory anyway.